Formatted code

+++ contracts/Flashswaptest.sol
@@ -2,11 +2,11 @@
 
 pragma solidity >=0.6.6 <0.8.0;
 
-import './UniswapV2Library.sol';
-import './interfaces/IERC20.sol';
-import './interfaces/IUniswapV2Pair.sol';
-import './interfaces/IUniswapV2Factory.sol';
-import './interfaces/IUniswapV2Router02.sol';
+import "./UniswapV2Library.sol";
+import "./interfaces/IERC20.sol";
+import "./interfaces/IUniswapV2Pair.sol";
+import "./interfaces/IUniswapV2Factory.sol";
+import "./interfaces/IUniswapV2Router02.sol";
 
 // @author Daniel Espendiller - https://github.com/Haehnchen/uniswap-arbitrage-flash-swap - espend.de
 //
@@ -21,7 +21,6 @@ import './interfaces/IUniswapV2Router02.sol';
 // e16:
 
 contract Flashswaptest {
-
     address public owner;
 
     constructor() {
@@ -30,32 +29,41 @@ contract Flashswaptest {
 
     // https://github.com/Haehnchen/uniswap-arbitrage-flash-swap/blob/main/contracts/Flashswap.sol
     function start(
-        uint _maxBlockNumber,
+        uint256 _maxBlockNumber,
         address _tokenPay, // source currency when we will get; example BNB
         address _tokenSwap, // swapped currency with the source currency; example BUSD
-        uint _amountTokenPay, // example: BNB => 10 * 1e18
+        uint256 _amountTokenPay, // example: BNB => 10 * 1e18
         address _sourceRouter,
         address _targetRouter,
         address _sourceFactory
     ) external {
-        require(block.number <= _maxBlockNumber, 'e00');
+        require(block.number <= _maxBlockNumber, "e00");
 
         // recheck for stopping and gas usage
-        (int profit, uint _tokenBorrowAmount) = check(_tokenPay, _tokenSwap, _amountTokenPay, _sourceRouter, _targetRouter);
+        (int256 profit, uint256 _tokenBorrowAmount) = check(
+            _tokenPay,
+            _tokenSwap,
+            _amountTokenPay,
+            _sourceRouter,
+            _targetRouter
+        );
         // profit needs to be greater than 0, otherwise revert a transaction
-        
+
         // comment out profit > 0, bypass this condition for testing purpose
         // require(profit > 0, 'e01');
 
         // https://docs.uniswap.org/protocol/V2/reference/smart-contracts/factory#getpair
-        address pairAddress = IUniswapV2Factory(_sourceFactory).getPair(_tokenPay, _tokenSwap);
-        require(pairAddress != address(0), 'e10');
+        address pairAddress = IUniswapV2Factory(_sourceFactory).getPair(
+            _tokenPay,
+            _tokenSwap
+        );
+        require(pairAddress != address(0), "e10");
 
         // scope for token{0,1}, avoids stack too deep errors
         address token0 = IUniswapV2Pair(pairAddress).token0();
         address token1 = IUniswapV2Pair(pairAddress).token1();
 
-        require(token0 != address(0) && token1 != address(0), 'e11');
+        require(token0 != address(0) && token1 != address(0), "e11");
 
         // https://docs.uniswap.org/protocol/V2/guides/smart-contract-integration/using-flash-swaps
         // _tokenPay can be zero because they should be sent from within a callback function
@@ -71,10 +79,10 @@ contract Flashswaptest {
     function check(
         address _tokenPay, // source currency when we will get; example BNB
         address _tokenSwap, // swapped currency with the source currency; example BUSD
-        uint _amountTokenPay, // example: BNB => 10 * 1e18
+        uint256 _amountTokenPay, // example: BNB => 10 * 1e18
         address _sourceRouter,
         address _targetRouter
-    ) public view returns(int, uint) {
+    ) public view returns (int256, uint256) {
         address[] memory path1 = new address[](2);
         address[] memory path2 = new address[](2);
 
@@ -83,23 +91,29 @@ contract Flashswaptest {
         // path2 represents the backward exchange from swapeed currency to source currency
         path1[1] = path2[0] = _tokenSwap;
 
-        uint amountOut = IUniswapV2Router02(_sourceRouter).getAmountsOut(_amountTokenPay, path1)[1];
-        uint amountRepay = IUniswapV2Router02(_targetRouter).getAmountsOut(amountOut, path2)[1];
+        uint256 amountOut = IUniswapV2Router02(_sourceRouter).getAmountsOut(
+            _amountTokenPay,
+            path1
+        )[1];
+        uint256 amountRepay = IUniswapV2Router02(_targetRouter).getAmountsOut(
+            amountOut,
+            path2
+        )[1];
 
         return (
-            int(amountRepay - _amountTokenPay), // our profit or loss; example output: BNB
+            int256(amountRepay - _amountTokenPay), // our profit or loss; example output: BNB
             amountOut // the amount we get from our input "_amountTokenPay"; example: BUSD amount
         );
     }
 
     function execute(
         address _sender,
-        uint _amount0,
-        uint _amount1,
+        uint256 _amount0,
+        uint256 _amount1,
         bytes calldata _data
     ) internal {
         // obtain an amount of token that you exchanged, for example BUSD
-        uint amountToken = _amount0 == 0 ? _amount1 : _amount0;
+        uint256 amountToken = _amount0 == 0 ? _amount1 : _amount0;
 
         IUniswapV2Pair iUniswapV2Pair = IUniswapV2Pair(msg.sender);
         address token0 = iUniswapV2Pair.token0();
@@ -122,8 +136,14 @@ contract Flashswaptest {
         // path1[1] = path2[0] = _amount0 == 0 ? token0 : token1;
         path1[1] = path2[0] = backward;
 
-        (address sourceRouter, address targetRouter) = abi.decode(_data, (address, address));
-        require(sourceRouter != address(0) && targetRouter != address(0), 'e12');
+        (address sourceRouter, address targetRouter) = abi.decode(
+            _data,
+            (address, address)
+        );
+        require(
+            sourceRouter != address(0) && targetRouter != address(0),
+            "e12"
+        );
 
         // IERC20 token that we will sell for otherToken, for example BUSD
         // IERC20 token = IERC20(_amount0 == 0 ? token1 : token0);
@@ -131,16 +151,20 @@ contract Flashswaptest {
         token.approve(targetRouter, amountToken);
 
         // calculate the amount of token how much input token should be reimbursed, BNB -> BUSD
-        uint amountRequired = IUniswapV2Router02(sourceRouter).getAmountsIn(amountToken, path2)[0];
+        uint256 amountRequired = IUniswapV2Router02(sourceRouter).getAmountsIn(
+            amountToken,
+            path2
+        )[0];
 
         // swap token and obtain equivalent otherToken amountRequired as a result, BUSD -> BNB
-        uint amountReceived = IUniswapV2Router02(targetRouter).swapExactTokensForTokens(
-            amountToken,
-            amountRequired, // we already now what we need at least for payback; get less is a fail; slippage can be done via - ((amountRequired * 19) / 981) + 1,
-            path1,
-            address(this), 
-            block.timestamp + 60
-        )[1];
+        uint256 amountReceived = IUniswapV2Router02(targetRouter)
+            .swapExactTokensForTokens(
+                amountToken,
+                amountRequired, // we already now what we need at least for payback; get less is a fail; slippage can be done via - ((amountRequired * 19) / 981) + 1,
+                path1,
+                address(this),
+                block.timestamp + 60
+            )[1];
 
         // fail if we didn't get enough tokens, bypass this condition for testing purpose
         // require(amountReceived > amountRequired, 'e13');
@@ -159,48 +183,92 @@ contract Flashswaptest {
 
     // c&p
     // pancake, pancakeV2, apeswap, kebab
-    function pancakeCall(address _sender, uint256 _amount0, uint256 _amount1, bytes calldata _data) external {
+    function pancakeCall(
+        address _sender,
+        uint256 _amount0,
+        uint256 _amount1,
+        bytes calldata _data
+    ) external {
         execute(_sender, _amount0, _amount1, _data);
     }
 
-    function waultSwapCall(address _sender, uint256 _amount0, uint256 _amount1, bytes calldata _data) external {
+    function waultSwapCall(
+        address _sender,
+        uint256 _amount0,
+        uint256 _amount1,
+        bytes calldata _data
+    ) external {
         execute(_sender, _amount0, _amount1, _data);
     }
 
-    function uniswapV2Call(address _sender, uint256 _amount0, uint256 _amount1, bytes calldata _data) external {
+    function uniswapV2Call(
+        address _sender,
+        uint256 _amount0,
+        uint256 _amount1,
+        bytes calldata _data
+    ) external {
         execute(_sender, _amount0, _amount1, _data);
     }
 
     // mdex
-    function swapV2Call(address _sender, uint256 _amount0, uint256 _amount1, bytes calldata _data) external {
+    function swapV2Call(
+        address _sender,
+        uint256 _amount0,
+        uint256 _amount1,
+        bytes calldata _data
+    ) external {
         execute(_sender, _amount0, _amount1, _data);
     }
 
     // pantherswap
-    function pantherCall(address _sender, uint256 _amount0, uint256 _amount1, bytes calldata _data) external {
+    function pantherCall(
+        address _sender,
+        uint256 _amount0,
+        uint256 _amount1,
+        bytes calldata _data
+    ) external {
         execute(_sender, _amount0, _amount1, _data);
     }
 
     // jetswap
-    function jetswapCall(address _sender, uint256 _amount0, uint256 _amount1, bytes calldata _data) external {
+    function jetswapCall(
+        address _sender,
+        uint256 _amount0,
+        uint256 _amount1,
+        bytes calldata _data
+    ) external {
         execute(_sender, _amount0, _amount1, _data);
     }
 
     // cafeswap
-    function cafeCall(address _sender, uint256 _amount0, uint256 _amount1, bytes calldata _data) external {
+    function cafeCall(
+        address _sender,
+        uint256 _amount0,
+        uint256 _amount1,
+        bytes calldata _data
+    ) external {
         execute(_sender, _amount0, _amount1, _data);
     }
 
     // @TODO: pending release
-    function BiswapCall(address _sender, uint256 _amount0, uint256 _amount1, bytes calldata _data) external {
+    function BiswapCall(
+        address _sender,
+        uint256 _amount0,
+        uint256 _amount1,
+        bytes calldata _data
+    ) external {
         execute(_sender, _amount0, _amount1, _data);
     }
 
     // @TODO: pending release
-    function wardenCall(address _sender, uint256 _amount0, uint256 _amount1, bytes calldata _data) external {
+    function wardenCall(
+        address _sender,
+        uint256 _amount0,
+        uint256 _amount1,
+        bytes calldata _data
+    ) external {
         execute(_sender, _amount0, _amount1, _data);
     }
 
     receive() external payable {}
 }
-
